# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: 伴AI API
  version: 1.27.0
  description: M1-M4 platform plus the M6 reliability control plane, governed Agent Studio, unified adjustable usage ledger, multi-instance runtime-configuration convergence, persisted cost and quality analytics with time trends, predictive budget burn and guarded model-switch recommendations, per-budget proactive forecast policies with sample gating, audited human recommendation feedback, revision-bound application and rollback-execution tracking, configurable before/after observation windows, downloadable periodic budget reviews, read-only rollback recommendations and post-rollback verification with acknowledged/closed human disposition, audited budget guardrails, automatic budget incidents with escalation plus email or management-console recovery notifications, baseline anomaly detection, Loki-backed redacted structured system logs with PostgreSQL fallback and Run/Trace correlation, threshold alert evaluation, downloadable incident evidence, durable W3C/OpenTelemetry Trace Context propagation across HTTP, Outbox, Kafka, Go/Python Agent workers and Langfuse, optional fail-open Langfuse LLM/Agent traces with generations, node observations and scores, operator DLQ replay, and compensation records. Every HTTP response includes canonical traceparent and X-Trace-ID correlation headers; clients may send traceparent/tracestate, while X-Trace-ID remains a normalized compatibility input.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/reliability:
    get:
      operationId: getReliability
      tags: [Platform]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Current degradation level, candidate hysteresis state, sampled signals, and effective policy.
          content: { application/json: { schema: { $ref: "#/components/schemas/ReliabilitySnapshot" } } }
...
  /v1/ops/logs:
    get:
      operationId: listOperatorSystemLogs
      tags: [Operations]
      security: [{ operatorBearerAuth: [] }]
      description: Queries Loki structured operational logs with explicit PostgreSQL fallback after server-side secret and content redaction. Prompt, response, request-body, credential, token, and cookie values are never returned.
      parameters:
        - { name: since, in: query, required: false, description: Defaults to one hour before until., schema: { type: string, format: date-time } }
        - { name: until, in: query, required: false, description: Defaults to the current time., schema: { type: string, format: date-time } }
        - { name: service, in: query, required: false, schema: { type: string, maxLength: 128 } }
        - { name: level, in: query, required: false, schema: { type: string, enum: [DEBUG, INFO, WARN, ERROR] } }
        - { name: q, in: query, required: false, description: Plain-text search over redacted event, message, error code, and bounded attributes., schema: { type: string, maxLength: 200 } }
        - { name: trace_id, in: query, required: false, description: Matches the canonical 32-hex W3C Trace ID shared by the request, Agent Run, logs, Kafka and Langfuse., schema: { type: string, pattern: "^[0-9a-f]{32}$" } }
        - { name: run_id, in: query, required: false, schema: { type: string, maxLength: 128 } }
        - { name: before_id, in: query, required: false, description: Loads records older than this log identifier., schema: { type: integer, minimum: 1 } }
        - { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 500, default: 100 } }
      responses:
        "200":
          description: Redacted logs, filtered summary, service facets, and an optional pagination cursor.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SystemLogPage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    operatorBearerAuth:
      type: http
      scheme: bearer
      description: Uses either legacy OPERATOR_TOKEN or an operator account token. When operator MFA is enabled, clients must also send X-Operator-TOTP.
  responses:
    Unauthorized:
      description: Authentication is missing, invalid, expired, or revoked.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    ValidationError:
      description: Request failed domain validation.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
  schemas:
    Error:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
...
    SystemLogEntry:
      type: object
      additionalProperties: false
      required: [id, occurred_at, service, environment, level, event, message, attributes]
      properties:
        id: { type: integer, minimum: 1 }
        occurred_at: { type: string, format: date-time }
        service: { type: string }
        environment: { type: string }
        level: { type: string, enum: [DEBUG, INFO, WARN, ERROR] }
        event: { type: string }
        message: { type: string, maxLength: 513 }
        trace_id: { type: string, pattern: "^[0-9a-f]{32}$" }
        span_id: { type: string, pattern: "^[0-9a-f]{16}$" }
        trace_flags: { type: string, pattern: "^[0-9a-f]{2}$" }
        run_id: { type: string }
        agent_trace_id: { type: string, pattern: "^[0-9a-f]{32}$", description: Langfuse Agent Trace ID; normally identical to trace_id, with a deterministic run_id-derived fallback for reconciliation without incoming context. }
        node: { type: string }
        error_code: { type: string }
        attributes: { type: object, additionalProperties: true, description: Redacted and size-bounded structured attributes. }
    SystemLogPage:
      type: object
      additionalProperties: false
      required: [items, summary, services, source]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/SystemLogEntry" } }
        next_before_id: { type: integer, minimum: 1 }
        summary:
          type: object
          additionalProperties: false
          required: [total, errors, warnings, services]
          properties:
            total: { type: integer, minimum: 0 }
            errors: { type: integer, minimum: 0 }
            warnings: { type: integer, minimum: 0 }
            services: { type: integer, minimum: 0 }
        services: { type: array, items: { type: string } }
        source: { type: string, enum: [loki, postgres, memory] }
        degraded: { type: boolean, description: True when a configured Loki query fell back to the durable local log store. }
        truncated: { type: boolean }
...
    ReliabilityPolicy:
      type: object
      additionalProperties: false
      required: [use_full_rag, extract_memory, preferred_model_class, long_skills_queued, accept_only]
      properties:
        use_full_rag: { type: boolean }
        extract_memory: { type: boolean }
        preferred_model_class: { type: string, enum: [primary, small, low_latency, local_only] }
        long_skills_queued: { type: boolean }
        accept_only: { type: boolean }
    ReliabilitySnapshot:
      type: object
      additionalProperties: false
      required: [level, reason, observed_at, changed_at, candidate_level, candidate_samples, queue_lag, oldest_job_age_ms, model_error_rate, p95_latency_ms, policy]
      properties:
        level: { type: string, enum: [L0, L1, L2, L3] }
        reason: { type: string }
        observed_at: { type: string, format: date-time }
        changed_at: { type: string, format: date-time }
        candidate_level: { type: string, enum: [L0, L1, L2, L3] }
        candidate_samples: { type: integer, minimum: 0 }
        queue_lag: { type: integer, minimum: 0 }
        oldest_job_age_ms: { type: integer, minimum: 0 }
        model_error_rate: { type: number, minimum: 0, maximum: 1 }
        p95_latency_ms: { type: integer, minimum: 0 }
        policy: { $ref: "#/components/schemas/ReliabilityPolicy" }
...